Conflicts:
	src/main/java/Employee.java

+++ src/main/java/Employee.java
@@ -7,7 +7,9 @@ public class Employee {
 	private String name;
 	private int age;
 
-	private double salary;
+
+	
+	private int salary;
 	private String address;
 	public String getName() {
 		return name;
